+++ SCCMCliCtrWPF/SCCMCliCtrWPF/Controls/WMIBrowser.xaml.cs
@@ -72,7 +72,7 @@ namespace ClientCenter.Controls
                 {
                     lClasses.Clear();
 
-                    List<PSObject> oClasses = oAgent.Client.GetObjectsFromPS("Get-WmiObject -Namespace '" + ((ListBoxItem)cb_Namespace.SelectedValue).Content.ToString() + "' -List | where { !$_.Name.StartsWith(\"__\") }", false);
+                    List<PSObject> oClasses = oAgent.Client.GetObjectsFromPS("Get-CimClass -Namespace '" + ((ListBoxItem)cb_Namespace.SelectedValue).Content.ToString() + "' | Where-Object { !$_.CimClassName.StartsWith(\"__\") } | Select-Object @{Name='Name';Expression={$_.CimClassName}}", false);
 
                     foreach (PSObject oClass in oClasses)
                     {
@@ -97,7 +97,7 @@ namespace ClientCenter.Controls
 
                     foreach (TreeViewItem tvi in tvObjects.Items)
                     {
-                        if (tvi.Header == cb_Classes.SelectedValue.ToString())
+                        if (string.Equals(tvi.Header as string, cb_Classes.SelectedValue.ToString(), StringComparison.Ordinal))
                         {
                             tvMain = tvi;
                             continue;
@@ -235,7 +235,7 @@ namespace ClientCenter.Controls
 
                     foreach (TreeViewItem tvi in tvObjects.Items)
                     {
-                        if (tvi.Header == sName)
+                        if (string.Equals(tvi.Header as string, sName, StringComparison.Ordinal))
                         {
                             tvMain = tvi;
                             continue;