+++ SCCMCliCtrWPF/SCCMCliCtrWPF/Helper/Converter.cs
@@ -56,7 +56,7 @@ namespace ClientCenter.Helper
 
         public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
         {
-            throw new NotImplementedException();
+            return Binding.DoNothing;
         }
     }
 }